@@ -11,13 +11,13 @@ import 'countries.dart';
 
 class CountryPickerDialog extends StatefulWidget {
   /// Callback that is called with selected Country
-  final ValueChanged<Country> onValuePicked;
+  final ValueChanged<Country>? onValuePicked;
 
   /// The (optional) title of the dialog is displayed in a large font at the top
   /// of the dialog.
   ///
   /// Typically a [Text] widget.
-  final Widget title;
+  final Widget? title;
 
   /// Padding around the title.
   ///
@@ -29,11 +29,11 @@ class CountryPickerDialog extends StatefulWidget {
   /// provided (but see [contentPadding]). If it _is_ null, then an extra 20
   /// pixels of bottom padding is added to separate the [title] from the
   /// [actions].
-  final EdgeInsetsGeometry titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Padding around the content.
 
-  final EdgeInsetsGeometry contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// The semantic label of the dialog used by accessibility frameworks to
   /// announce screen transitions when the dialog is opened and closed.
@@ -46,20 +46,20 @@ class CountryPickerDialog extends StatefulWidget {
   ///
   ///  * [SemanticsConfiguration.isRouteName], for a description of how this
   ///    value is used.
-  final String semanticLabel;
+  final String? semanticLabel;
 
   /// Filters the available country list
-  final ItemFilter itemFilter;
+  final ItemFilter? itemFilter;
 
   /// [Comparator] to be used in sort of country list
-  final Comparator<Country> sortComparator;
+  final Comparator<Country>? sortComparator;
 
   /// List of countries that are placed on top
-  final List<Country> priorityList;
+  final List<Country>? priorityList;
 
   ///Callback that is called with selected item of type Country which returns a
   ///Widget to build list view item inside dialog
-  final ItemBuilder itemBuilder;
+  final ItemBuilder? itemBuilder;
 
   /// The (optional) horizontal separator used between title, content and
   /// actions.
@@ -76,23 +76,23 @@ class CountryPickerDialog extends StatefulWidget {
   final bool isSearchable;
 
   /// The optional [decoration] of search [TextField]
-  final InputDecoration searchInputDecoration;
+  final InputDecoration? searchInputDecoration;
 
   ///The optional [cursorColor] of search [TextField]
-  final Color searchCursorColor;
+  final Color? searchCursorColor;
 
   ///The search empty view is displayed if nothing returns from search result
-  final Widget searchEmptyView;
+  final Widget? searchEmptyView;
 
   ///By default the dialog will be popped of the navigator on selection of a value.
   ///Set popOnPick to false to prevent this behaviour.
   final bool popOnPick;
 
   ///Filters the country list for search
-  final SearchFilter searchFilter;
+  final SearchFilter? searchFilter;
 
   CountryPickerDialog({
-    Key key,
+    Key? key,
     this.onValuePicked,
     this.title,
     this.titlePadding,
@@ -121,9 +121,9 @@ class CountryPickerDialog extends StatefulWidget {
 }
 
 class SingleChoiceDialogState extends State<CountryPickerDialog> {
-  List<Country> _allCountries;
+  late List<Country> _allCountries;
 
-  List<Country> _filteredCountries;
+  late List<Country> _filteredCountries;
 
   @override
   void initState() {
@@ -135,9 +135,9 @@ class SingleChoiceDialogState extends State<CountryPickerDialog> {
     }
 
     if (widget.priorityList != null) {
-      widget.priorityList.forEach((Country country) => _allCountries
+      widget.priorityList!.forEach((Country country) => _allCountries
           .removeWhere((Country c) => country.isoCode == c.isoCode));
-      _allCountries.insertAll(0, widget.priorityList);
+      _allCountries.insertAll(0, widget.priorityList!);
     }
 
     _filteredCountries = _allCountries;
@@ -149,7 +149,7 @@ class SingleChoiceDialogState extends State<CountryPickerDialog> {
   Widget build(BuildContext context) {
     return MyAlertDialog(
       title: _buildHeader(),
-      contentPadding: widget.contentPadding,
+      contentPadding: widget.contentPadding!,
       semanticLabel: widget.semanticLabel,
       content: _buildContent(context),
       isDividerEnabled: widget.isDividerEnabled,
@@ -164,10 +164,10 @@ class SingleChoiceDialogState extends State<CountryPickerDialog> {
             children: _filteredCountries
                 .map((item) => SimpleDialogOption(
                       child: widget.itemBuilder != null
-                          ? widget.itemBuilder(item)
-                          : Text(item.name),
+                          ? widget.itemBuilder!(item)
+                          : Text(item.name!),
                       onPressed: () {
-                        widget.onValuePicked(item);
+                        widget.onValuePicked!(item);
                         if (widget.popOnPick) {
                           Navigator.pop(context);
                         }
@@ -195,7 +195,7 @@ class SingleChoiceDialogState extends State<CountryPickerDialog> {
   _buildTitle() {
     return widget.titlePadding != null
         ? Padding(
-            padding: widget.titlePadding,
+            padding: widget.titlePadding!,
             child: widget.title,
           )
         : widget.title;
@@ -210,17 +210,17 @@ class SingleChoiceDialogState extends State<CountryPickerDialog> {
         setState(() {
           _filteredCountries = _allCountries
               .where((Country country) => widget.searchFilter == null
-                  ? country.name
+                  ? country.name!
                           .toLowerCase()
                           .startsWith(value.toLowerCase()) ||
-                      country.phoneCode.startsWith(value.toLowerCase()) ||
-                      country.isoCode
+                      country.phoneCode!.startsWith(value.toLowerCase()) ||
+                      country.isoCode!
                           .toLowerCase()
                           .startsWith(value.toLowerCase()) ||
-                      country.iso3Code
+                      country.iso3Code!
                           .toLowerCase()
                           .startsWith(value.toLowerCase())
-                  : widget.searchFilter(country, value))
+                  : widget.searchFilter!(country, value))
               .toList();
         });
       },
